refactor: split audio and render support from workflow runner
